implemented generate_piece,, get_random_color, get_random_shape, get_random_number and mod
This code will draw one square in the top left corner of the screen. The square does not move, and it is the same square each time because the random functions give the same value each time. Also, the coordinates of the pieces to be drawn need to be set, as they were all set to zero.